requirements.txt
altair
pandas
streamlit
tensorflow
matplotlib
numpy
opencv-python-headless
torch
albumentations==1.0.3
torchvision
pillow
h5py
torchvision
